requirements.txt
psutil>=5.9.0
rich>=13.0.0
streamlit>=1.30.0
plotly>=5.18.0
pyyaml>=6.0